feat: Add functions for managing stream readers in Turbo-restler
This commit adds new functions to the Turbo-restler library for managing stream readers. The following functions have been added:
- `func (ws *WebStream) SetHandler(handler WsHandler) *WebStream`
- `func (ws *WebStream) SetErrHandler(errHandler ErrHandler) *WebStream`
- `func (ws *WebStream) AddSubscriptions(handlerId string, handler WsHandler)`
- `func (ws *WebStream) RemoveSubscriptions(handlerId string)`

These functions allow users to set handlers and error handlers for stream readers, as well as add and remove subscriptions. The commit also includes an example usage in the `main` function of the `web_stream` package.