chore: Update WebStream Subscribe method to support multiple subscriptions
This commit modifies the `Subscribe` method in the `WebStream` struct to accept multiple subscriptions as variadic parameters. It checks if there are any subscriptions provided and sends a subscription request for each one. This change improves the flexibility and usability of the `Subscribe` method.

Note: This commit is based on recent user commits and repository commits.